Refactoring testHelper into relevant files
