+++ cypress/integration/headerSpec.js
@@ -2,7 +2,7 @@ import {
   getElement,
   shouldContainText,
   shouldContainStyles,
-} from '../support/testHelper';
+} from '../support/bodyTestHelper';
 
 describe('Header Tests', () => {
   // eslint-disable-next-line no-undef
