+++ cypress/integration/metaNewsSpec.js
@@ -1,10 +1,9 @@
+import { getElement, getSecondElement } from '../support/bodyTestHelper';
 import {
-  getElement,
-  getSecondElement,
   openGraphMeta,
   retrieveMetaDataContent,
   twitterMeta,
-} from '../support/testHelper';
+} from '../support/metaTestHelper';
 
 describe('Article Meta Tests', () => {
   // eslint-disable-next-line no-undef
