+++ cypress/integration/metaPersianSpec.js
@@ -1,9 +1,9 @@
+import { getElement } from '../support/bodyTestHelper';
 import {
-  getElement,
   openGraphMeta,
   retrieveMetaDataContent,
   twitterMeta,
-} from '../support/testHelper';
+} from '../support/metaTestHelper';
 
 describe('Persian Article Meta Tests', () => {
   // eslint-disable-next-line no-undef
