+++ cypress/integration/statusSpec.js
@@ -1,4 +1,4 @@
-import { testNonHTMLResponseCode } from '../support/testHelper';
+import { testNonHTMLResponseCode } from '../support/metaTestHelper';
 
 describe('Simorgh Status', () => {
   it('should return 200', () => {
